Migrate Message to code_builder
The String `toJsonSuffix` is replaced by a
`Expression Function(Expression)`. This should make it a little nicer to
refactor when adding support for Map fields (and was the main purpose
for migrating to code_builder).

Add `hash_code.dart` to pull out implementation of hashCode Methods.

Add `equality.dart` to pull out implementation of operator== Methods

Add `fields.dart` to pull out MessageField and related classes.